changed cleanct.go to ctmonitor.go.  Cleaned up code for code review.
